[ACPO] Initial commit with core supporting work
This initial change delivers an example unrolling model together with
many scripts used for inference and quality analysis. It also comprises
python script to run as parallel processes for a compiler to do
inference via named pipes, for which the setup has been made. A
companion set of changes to invoke these scripts is checked into the
LLVM-based compiler framework.